Correctly support fractional plural parts in tofu, soysauce and soy py.
Prior to this change:
 * SoySauce would reject all floating point values at runtime with a class cast exception
 * Tofu would support floating point values in english but would truncate floating point values to integers in non-default locales.
 * SoyPy would also truncate values to match cases.

With this CL they all behave identically for basic usage.  The plural implementation in soy_py is fairly limited (doesn't support advanced plural rules), so there are likely still issues there.

GITHUB_BREAKING_CHANGES=server side rendering will now match JavaScript behavior for fractional plurals.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=245946307

+++ java/src/com/google/template/soy/msgs/restricted/SoyMsgPluralPart.java
@@ -76,7 +76,7 @@ public final class SoyMsgPluralPart extends SoyMsgPart {
    *     is known that there are no non-specific plural parts (This is commonly the case for default
    *     messages, since soy only allows direct specification of explicit or 'other').
    */
-  public ImmutableList<SoyMsgPart> lookupCase(long pluralValue, @Nullable ULocale locale) {
+  public ImmutableList<SoyMsgPart> lookupCase(double pluralValue, @Nullable ULocale locale) {
     // TODO(lukes): clean up this method, the control flow is overly complex.  It could be cleaned
     // up by separating cases into 3 different lists.
